requirements.txt
numpy
pygame
neat-python